Linear Regression(Fuel Consunption)
linear regression on the "FuelConsumption" dataset. It begins with data loading and selecting specific features for analysis. Visualization includes histograms and scatter plots to understand feature distributions and relationships. Data is split into training and testing sets, with 80% for training and 20% for testing. A linear regression model is trained using the training data. The model's performance is evaluated on the test data, assessing metrics like mean absolute error, mean squared error, and R-squared score.